Separate skip-based util methods
